feat(site-builder): improve passing of context to walrus binary
