+++ site-builder/src/display.rs
@@ -49,13 +49,3 @@ pub fn done() {
         .unwrap();
     }
 }
-
-pub fn warn<S: Display>(message: S) {
-    if cfg!(not(test)) {
-        crossterm::execute!(
-            stdout(),
-            Print(format!("\n{} {message}\n", "Warning:".yellow())),
-        )
-        .unwrap();
-    }
-}
