+++ sites-config.yaml
@@ -3,7 +3,8 @@ contexts:
     # module: site
     # portal: wal.app
     package: 0xf99aee9f21493e1590e7e5a9aea6f343a1f381031a04a732724871fc294be799
-    # general:
+    general:
+        wallet_env: testnet
     #   rpc_url: https://fullnode.testnet.sui.io:443
     #   wallet: /path/to/.sui/sui_config/client.yaml
     #   walrus_binary: /path/to/walrus
@@ -13,7 +14,8 @@ contexts:
     # module: site
     # portal: wal.app
     package: 0x26eb7ee8688da02c5f671679524e379f0b837a12f1d1d799f255b7eea260ad27
-    # general:
+    general:
+        wallet_env: mainnet
     #   rpc_url: https://fullnode.mainnet.sui.io:443
     #   wallet: /path/to/.sui/sui_config/client.yaml
     #   walrus_binary: /path/to/walrus
